# Conflicts:
#   FRONT/muscle_maker/src/app/(beforeLogin)/_component/Main.tsx

Former-commit-id: 1c594eed418b692193ac19738691afdfd1c0091d

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/clubs/request/ClubTransitRequest.java
@@ -7,6 +7,6 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class ClubTransitRequest {
-    private long clubId;
-    private long userId;
+    private int clubId;
+    private int userId;
 }
