+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/feeds/request/FeedDeleteRequest.java
@@ -10,5 +10,5 @@ import lombok.*;
 @Builder
 public class FeedDeleteRequest {
     private  String Authorization;
-    private  Long feedId;
+    private  int feedId;
 }
